requirements.txt
Django~=5.0.2
psycopg2-binary~=2.9.9
python-dotenv==1.0.1
djangorestframework==3.14.0
gunicorn>=20.1.0